Refactor schema module
